Attempt to sync ropes
In addition to making sure that the ropes are spinning at the proper
speed, we make sure that they are in sync with each other.

To do this we check if one of the ropes reached the button significantly
before the other, and if so give a temporary speed boost - not large
enough to affect the RPM of the rope, but enough to close the distance.

It works reasonably well. It is important to make sure that the total
speeds are less (with a reasonable buffer) than 100 or bad things might
happen...